refactor(app): operation-log 域内聚至 features
logStore、logService、types 迁入 features/operation-log；清空根级 services/ 与 types/log。
